Moving request verification test helper function up.
