+++ cmd/config_test.go
@@ -14,28 +14,6 @@ import (
 	"github.com/onsi/gomega/ghttp"
 )
 
-func appendPostHandler(server *ghttp.Server, authToken string, statusCode int, expectedRequestJson string, responseBody string) {
-	server.AppendHandlers(
-		ghttp.CombineHandlers(
-			ghttp.VerifyRequest("POST", "/"),
-			ghttp.VerifyHeader(http.Header{
-				"Authorization": []string{authToken},
-			}),
-			ghttp.VerifyContentType("application/json; charset=utf-8"),
-			// From Gomegas ghttp.VerifyJson to avoid the
-			// VerifyContentType("application/json") check
-			// that fails with "application/json; charset=utf-8"
-			func(w http.ResponseWriter, req *http.Request) {
-				body, err := ioutil.ReadAll(req.Body)
-				req.Body.Close()
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(body).Should(MatchJSON(expectedRequestJson), "JSON Mismatch")
-			},
-			ghttp.RespondWith(statusCode, `{ "data": `+responseBody+`}`),
-		),
-	)
-}
-
 type configYaml struct {
 	TempHome string
 	Path     string
